refactor: extract map runner policy resolution
